Case history updates (#18)

+++ src/main/java/uk/gov/ons/census/action/model/entity/Case.java
@@ -15,7 +15,7 @@ import lombok.Data;
     })
 public class Case {
 
-  @Id private long caseRef;
+  @Id private int caseRef;
 
   @Column(name = "case_id")
   private UUID caseId;
